+++ Demo/iOS/EditorScreen.swift
@@ -59,7 +59,7 @@ private extension EditorScreen {
             trailingButtons: {}
         ) {
             var sheet = $0
-            sheet.colorPickers = .all
+            // sheet.colorPickers = .all
             return sheet
         }
     }